requirements.txt
Django>=4.2.7
mysqlclient>=2.2.0
Pillow>=10.1.0
python-dotenv>=1.0.0
stripe>=7.2.0
django-crispy-forms>=2.1
whitenoise>=6.6.0
